pyproject.toml
[project]
name = "legislation-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "datasets>=3.4.1",
    "docopt>=0.6.2",
    "markitdown>=0.0.1a3",
    "pip>=24.3.1",
    "psutil>=7.0.0",
    "pymupdf>=1.25.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "selenium>=4.29.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.12.2",
    "unidecode>=1.3.8",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.0",
]
